+++ api/app/config.py
@@ -1,5 +1,6 @@
 from dotenv import load_dotenv
 import os, sys
+from rdflib import Namespace
 
 load_dotenv()
 TRIPLESTORE_URL = os.environ.get("TRIPLESTORE_URL", None)
@@ -15,3 +16,8 @@ UPDATE_URL = f"{TRIPLESTORE_URL}/{DATASET_NAME}/update"
 
 JWKS_URL = os.environ.get("JWKS_URL", None)
 JWT_AUD = os.environ.get("JWT_AUD", None)
+
+cddo_graph = Namespace("http://marketplace.cddo.gov.uk/graph/")
+ASSET_GRAPH = cddo_graph.assets
+USER_GRAPH = cddo_graph.users
+SHARE_GRAPH = cddo_graph.shares
