--- api/app/db.py
@@ -1,102 +1,0 @@
-import json
-from datetime import datetime
-from . import utils
-from . import model as m
-from . import sparql
-
-
-def _resolve_media_type_label(db_result_dict):
-    if "mediaTypeLabel" in db_result_dict:
-        db_result_dict["mediaType"] = db_result_dict["mediaTypeLabel"]
-        db_result_dict.pop("mediaTypeLabel")
-
-
-def search(q: str = ""):
-    if q == "":
-        q = "*"
-
-    # TODO What else do we need to do to sanitise the query string?
-    q = utils.sanitise_search_query(q)
-    query_results = sparql.run_query("asset_search.sparql", q=q)
-    for r in query_results:
-        _resolve_media_type_label(r)
-    result_dicts = sparql.aggregate_query_results_by_key(
-        query_results, group_key="identifier"
-    )
-    result_dicts = [utils.enrich_query_result_dict(r) for r in result_dicts]
-    result_dicts = [utils.munge_asset_summary_response(r) for r in result_dicts]
-
-    return result_dicts
-
-
-def fetch_distribution_details(distribution_ids):
-    distribution_ids = [f"<{i}>" for i in distribution_ids]
-    results = sparql.run_query(
-        "distribution_detail.sparql", distribution=distribution_ids
-    )
-    for r in results:
-        _resolve_media_type_label(r)
-    return sparql.aggregate_query_results_by_key(results, group_key="distribution")
-
-
-def asset_detail(asset_id: str):
-    query_results = sparql.run_query("asset_detail.sparql", asset_id=asset_id)
-    asset_result_dicts = sparql.aggregate_query_results_by_key(
-        query_results, group_key="resourceUri"
-    )
-    result_dicts = [utils.enrich_query_result_dict(r) for r in asset_result_dicts]
-    assert len(result_dicts) == 1
-    asset = result_dicts[0]
-
-    asset["identifier"] = asset_id
-
-    contactPoint = {
-        "name": asset["contactName"],
-        "email": asset["contactEmail"],
-        "contactTelephone": asset.get("contentTelephone", None),
-        "contactAddress": asset.get("contentAddress", None),
-    }
-
-    asset["contactPoint"] = m.ContactPoint.model_validate(contactPoint)
-
-    if asset["type"] == m.assetType.dataset:
-        distributions = [
-            utils.remove_keys(r, ["distribution"])
-            for r in fetch_distribution_details(asset["distribution"])
-        ]
-        asset["distributions"] = [
-            m.DistributionSummary.model_validate(d) for d in distributions
-        ]
-    return asset
-
-
-def new_user(user_id: str, user_email: str):
-    query_results = sparql.run_update(
-        "new_user.sparql", user_id=user_id, user_email=user_email
-    )
-    return query_results
-
-
-def get_user_by_id(user_id: str):
-    query_results = sparql.run_query("get_user_by_id.sparql", user_id=user_id)
-    return query_results
-
-
-def get_share_request_forms(user_id: str):
-    query_results = sparql.run_query("get_share_request_forms.sparql", user_id=user_id)
-    forms = {r["assetId"]: r["sharedata"] for r in query_results}
-    return forms
-
-
-def upsert_sharedata(user_id: str, form_str: str):
-    form_dict = json.loads(form_str)
-    query_results = sparql.run_update(
-        "update_share_request_form.sparql",
-        id=form_dict["requestId"],
-        user_id=user_id,
-        asset_id=form_dict["dataAsset"],
-        sharedata=form_str,
-        current_time=datetime.now().isoformat(),
-        status=form_dict["status"],
-    )
-    return query_results
