+++ api/queries/distribution_detail.sparql
@@ -1,11 +1,9 @@
 PREFIX dcat: <http://www.w3.org/ns/dcat#>
 PREFIX dct: <http://purl.org/dc/terms/>
 PREFIX rdfs: <http://www.w3.org/2000/01/rdf-schema#>
-PREFIX cddo_graph: <http://marketplace.cddo.gov.uk/graph/>
 PREFIX skos: <http://www.w3.org/2004/02/skos/core#> 
 
 SELECT ?distribution ?title ?modified ?mediaType ?identifier ?license ?issued ?byteSize ?mediaTypeLabel
-FROM cddo_graph:assets
 WHERE {{
     VALUES ?distribution { $distribution }
     ?distribution dct:title ?title ;
