+++ api/queries/get_share_request_forms.sparql
@@ -1,8 +1,6 @@
 PREFIX cddo: <http://marketplace.cddo.gov.uk#>
-PREFIX cddo_graph: <http://marketplace.cddo.gov.uk/graph/>
 
 SELECT ?assetId ?sharedata
-FROM cddo_graph:shares
 WHERE {
   ?share cddo:user "$user_id" ;
          cddo:asset ?assetId ;
