+++ api/queries/get_user_by_id.sparql
@@ -2,10 +2,8 @@ PREFIX dct: <http://purl.org/dc/terms/>
 PREFIX vcard: <http://www.w3.org/2006/vcard/ns#>
 PREFIX cddo: <http://marketplace.cddo.gov.uk#>
 PREFIX cddo_user: <http://marketplace.cddo.gov.uk/user/>
-PREFIX cddo_graph: <http://marketplace.cddo.gov.uk/graph/>
 
 SELECT ?email
-FROM cddo_graph:users
 WHERE {
     cddo_user:$user_id vcard:hasEmail ?email
 }
