+++ api/queries/update_share_request_form.sparql
@@ -4,16 +4,13 @@ PREFIX vcard: <http://www.w3.org/2006/vcard/ns#>
 PREFIX cddo: <http://marketplace.cddo.gov.uk#>
 PREFIX cddo_user: <http://marketplace.cddo.gov.uk/user/>
 PREFIX cddo_share: <http://marketplace.cddo.gov.uk/share/>
-PREFIX cddo_graph: <http://marketplace.cddo.gov.uk/graph/>
 
 INSERT DATA
 { 
-    GRAPH cddo_graph:shares {
-        cddo_share:$id dct:identifier "$id";
-                       cddo:user "$user_id";
-                       cddo:asset "$asset_id";
-                       cddo:sharedata '$sharedata';
-                       dct:modified "$current_time";
-                       adms:status "$status" .
-    }
+      cddo_share:$id dct:identifier "$id";
+                      cddo:user "$user_id";
+                      cddo:asset "$asset_id";
+                      cddo:sharedata '$sharedata';
+                      dct:modified "$current_time";
+                      adms:status "$status" .
 }
